# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Backend App Ecommerce API",
        "description": "Documentación de la API del Ecommerce",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:3001",
            "description": "Servidor Local"
        }
    ],
    "paths": {
...
        "/api/auth/login": {
            "post": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Iniciar sesión",
                "operationId": "8dcb70df1020986038d098cc08d05dae",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "correo",
                                    "password"
                                ],
                                "properties": {
                                    "correo": {
                                        "type": "string",
                                        "example": "admin@correo.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Token generado correctamente"
                    },
                    "401": {
                        "description": "Credenciales no válidas"
                    },
                    "500": {
                        "description": "Error al procesar el token"
                    }
                }
            }
        },
        "/api/auth/register": {
            "post": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Registrar nuevo cliente",
                "description": "Endpoint público para que los clientes se registren en la tienda virtual completando todos sus datos. Automáticamente asigna el rol de USUARIO EXTERNO y devuelve el token de sesión.",
                "operationId": "d764dd091cc4494ae0baf360b03319f3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombres",
                                    "apellido_paterno",
                                    "correo",
                                    "password"
                                ],
                                "properties": {
                                    "tipo_documento_identidad_id": {
                                        "type": "integer",
                                        "example": 1,
                                        "nullable": true
                                    },
                                    "numero_documento": {
                                        "type": "string",
                                        "example": "12345678",
                                        "nullable": true
                                    },
                                    "nombres": {
                                        "type": "string",
                                        "example": "Juan Carlos"
                                    },
                                    "apellido_paterno": {
                                        "type": "string",
                                        "example": "Perez"
                                    },
                                    "apellido_materno": {
                                        "type": "string",
                                        "example": "Gomez",
                                        "nullable": true
                                    },
                                    "numero_celular": {
                                        "type": "string",
                                        "example": "999888777",
                                        "nullable": true
                                    },
                                    "departamento_id": {
                                        "type": "integer",
                                        "example": 15,
                                        "nullable": true
                                    },
                                    "provincia_id": {
                                        "type": "integer",
                                        "example": 1501,
                                        "nullable": true
                                    },
                                    "distrito_id": {
                                        "type": "integer",
                                        "example": 150101,
                                        "nullable": true
                                    },
                                    "fecha_nacimiento": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "1990-05-15",
                                        "nullable": true
                                    },
                                    "genero_id": {
                                        "type": "integer",
                                        "example": 1,
                                        "nullable": true
                                    },
                                    "correo": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "juan@correo.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "123456"
                                    },
                                    "acepto_termino_condiciones": {
                                        "type": "integer",
                                        "example": 1,
                                        "enum": [
                                            0,
                                            1
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Usuario registrado y autenticado correctamente"
                    },
                    "422": {
                        "description": "Errores de validación"
                    }
                }
            }
        },
        "/api/auth/me": {
            "get": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Obtener usuario autenticado",
                "operationId": "b9abb1a7a74670a19c215c2c133f14d8",
                "responses": {
                    "200": {
                        "description": "Información del usuario"
                    },
                    "401": {
                        "description": "No autorizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/auth/logout": {
            "post": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Cerrar sesión",
                "description": "Invalida el token JWT del usuario autenticado.",
                "operationId": "69281b12abb272c76871f19cb17ca563",
                "responses": {
                    "200": {
                        "description": "Sesión cerrada correctamente"
                    },
                    "401": {
                        "description": "Token inválido o no autorizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/auth/refresh": {
            "post": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Renovar token",
                "description": "Genera un nuevo token JWT para el usuario autenticado.",
                "operationId": "b2eae7a8032954d25b46c7ec6f489c68",
                "responses": {
                    "200": {
                        "description": "Nuevo token generado correctamente"
                    },
                    "401": {
                        "description": "Token inválido o expirado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Ingrese el token JWT con el formato: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}